restructured service interfaces to avoid duplicate code
