+++ src/main/java/com/globallogic/amcr/service/contactcomponent/FeedbackService.java
@@ -1,40 +1,32 @@
 package com.globallogic.amcr.service.contactcomponent;
 
-import com.globallogic.amcr.persistence.model.contactcomponent.*;
-import com.globallogic.amcr.persistence.model.contactcomponent.FeedbackAttachment;
-import com.globallogic.amcr.persistence.model.contactcomponent.FeedbackAttachmentResponse;
+import com.globallogic.amcr.model.contactcomponent.Feedback;
+import com.globallogic.amcr.model.contactcomponent.FeedbackAttachment;
+import com.globallogic.amcr.model.contactcomponent.FeedbackAttachmentResponse;
+import com.globallogic.amcr.service.Service;
 
 import java.util.List;
 import java.util.UUID;
 
-public interface FeedbackService {
+public interface FeedbackService extends Service<Feedback> {
+    // Extra save and get methods
 
-    /**
-     * @param feedback           the feedback object with the data to be saved
-     * @param feedbackAttachment the feedbackAttachment that needs to be saved belonging to the feedback
-     * @return a response entity with the status of the request
+    /** Sends a Feedback object to the FeedbackDao and a FeedbackAttachment object to the FeedbackAttachmentDao
+     *
+     * @param feedback The feedback object to be saved
+     * @param feedbackAttachment The feedback attachment object to be saved
+     * @return Returns the complete Feedback object as saved to the database
      */
     Feedback save(Feedback feedback, FeedbackAttachment feedbackAttachment);
 
     /**
-     * @param id the ID of the requested feedback entry
-     * @return returns the requested entry
-     */
-    Feedback get(UUID id);
-
-    /**
-     * @return a list of all entries in the feedback that also contains download links to attachments if applicable
-     */
-    List<Feedback> getAll();
-
-    /**
-     * @return a list of the 10 newest entries in the feedback table
+     * @return Returns a list of the 10 newest entries in the feedback table
      */
     List<Feedback> getLatest();
 
     /**
-     * @param last the 'order' number of the last returned feedback entry
-     * @return a list of 10 feedback entries that come after the 'last' entry
+     * @param last The 'order' number of the last returned feedback entry
+     * @return Returns a list of 10 feedback entries that come after the 'last' entry
      */
     List<Feedback> getOlder(int last);
