+++ src/main/java/com/globallogic/amcr/service/impl/contactcomponent/FeedbackServiceImpl.java
@@ -1,0 +1,99 @@
+package com.globallogic.amcr.service.impl.contactcomponent;
+
+import com.globallogic.amcr.controller.casestudies.CaseStudyController;
+import com.globallogic.amcr.model.contactcomponent.Feedback;
+import com.globallogic.amcr.model.contactcomponent.FeedbackAttachment;
+import com.globallogic.amcr.model.contactcomponent.FeedbackAttachmentResponse;
+import com.globallogic.amcr.repository.contactcomponent.FeedbackAttachmentDao;
+import com.globallogic.amcr.repository.contactcomponent.FeedbackDao;
+import com.globallogic.amcr.service.contactcomponent.EmailService;
+import com.globallogic.amcr.service.contactcomponent.FeedbackService;
+import com.globallogic.amcr.utils.Assert;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
+import java.util.List;
+import java.util.UUID;
+
+@Service
+public class FeedbackServiceImpl implements FeedbackService {
+    private final Logger Log = LoggerFactory.getLogger(CaseStudyController.class);
+    private final FeedbackDao feedbackDao;
+    private final FeedbackAttachmentDao feedbackAttachmentDao;
+    private final EmailService emailService;
+
+    public FeedbackServiceImpl(FeedbackDao feedbackDao, FeedbackAttachmentDao feedbackAttachmentDao, EmailService emailService) {
+        this.feedbackDao = Assert.assertNull(feedbackDao, "Feedback DAO cannot be null");
+        this.feedbackAttachmentDao = Assert.assertNull(feedbackAttachmentDao, "File DAO cannot be null");
+        this.emailService = Assert.assertNull(emailService, "Email service cannot be null");
+    }
+
+    @Override
+    @Transactional
+    public Feedback save(Feedback feedback) {
+        Assert.assertNull(feedback, "Feedback cannot be null");
+        UUID feedbackId = UUID.randomUUID();
+        Log.debug("Service saving feedback");
+        Feedback createdFeedback = feedbackDao.save(feedback, feedbackId);
+        Log.debug("Service sending email");
+        emailService.sendMail(feedback, feedbackId);
+        return createdFeedback;
+
+    }
+
+    @Override
+    @Transactional
+    public Feedback save(Feedback feedback, FeedbackAttachment feedbackAttachment) {
+        Assert.assertNull(feedback, "Feedback cannot be null");
+        Assert.assertNull(feedbackAttachment, "Feedback attachment cannot be null");
+        UUID feedbackId = UUID.randomUUID();
+        Log.debug("Service saving feedback");
+        Feedback createdFeedback = feedbackDao.save(feedback, feedbackId);
+        Log.debug("Service saving feedback attachment");
+        feedbackAttachmentDao.save(feedbackAttachment, feedbackId);
+        Log.debug("Service sending email");
+        emailService.sendMail(feedback, feedbackId);
+        return createdFeedback;
+
+    }
+
+    @Override
+    @Transactional
+    public Feedback get(UUID id) {
+        Assert.assertNull(id, "ID cannot be null to request feedback");
+        Log.debug("Service requesting feedback with ID {}", id);
+        return feedbackDao.get(id);
+    }
+
+    @Override
+    @Transactional(readOnly = true)
+    public List<Feedback> getAll() {
+        Log.debug("Service requesting all feedback entries");
+        return feedbackDao.getAll();
+    }
+
+    @Override
+    @Transactional(readOnly = true)
+    public List<Feedback> getLatest() {
+        Log.debug("Service requesting latest feedback");
+        return feedbackDao.getLatest();
+    }
+
+    @Override
+    @Transactional(readOnly = true)
+    public List<Feedback> getOlder(int last) {
+        Assert.assertNull(last, "Last cannot be null");
+        Log.debug("Service requesting 10 older feedback entries");
+        return feedbackDao.getOlder(last);
+    }
+
+    @Override
+    @Transactional(readOnly = true)
+    public FeedbackAttachmentResponse getAttachment(UUID id) {
+        Assert.assertNull(id, "ID cannot be null for file request");
+        Log.debug("Service requesting attachment with ID {}", id);
+        return feedbackAttachmentDao.get(id);
+    }
+}
